PLFM-3260

+++ integration-test/src/test/java/org/sagebionetworks/IT930ActivityController.java
@@ -61,9 +61,11 @@ public class IT930ActivityController {
 	
 	@AfterClass
 	public static void afterClass() throws Exception {
-		adminSynapse.deleteUser(userToDelete);
+		if (userToDelete != null) {
+			adminSynapse.deleteUser(userToDelete);
+		}
 	}
-		
+
 	@Test 
 	public void testActivityCrud() throws Exception {
 		// create
